+++ dflib/src/main/java/com/nhl/dflib/series/BooleanBaseSeries.java
@@ -4,9 +4,11 @@ import com.nhl.dflib.BooleanSeries;
 import com.nhl.dflib.DataFrame;
 import com.nhl.dflib.IntSeries;
 import com.nhl.dflib.Series;
+import com.nhl.dflib.SeriesGroupBy;
 import com.nhl.dflib.ValueMapper;
 import com.nhl.dflib.ValuePredicate;
 import com.nhl.dflib.concat.SeriesConcat;
+import com.nhl.dflib.groupby.SeriesGrouper;
 import com.nhl.dflib.series.builder.BooleanAccumulator;
 import com.nhl.dflib.series.builder.IntAccumulator;
 import com.nhl.dflib.series.builder.ObjectAccumulator;
@@ -431,6 +433,18 @@ public abstract class BooleanBaseSeries implements BooleanSeries {
     @Override
     public DataFrame valueCounts() {
         return ValueCounts.valueCountsNoNulls(this);
+    }
+
+    // TODO: some optimized version of "primitive" group by ...
+
+    @Override
+    public SeriesGroupBy<Boolean> group() {
+        return group(i -> i);
+    }
+
+    @Override
+    public SeriesGroupBy<Boolean> group(ValueMapper<Boolean, ?> by) {
+        return new SeriesGrouper<>(by).group(this);
     }
 
     @Override