+++ dflib/src/test/java/com/nhl/dflib/series/ByRowSeries_HeadTest.java
@@ -1,7 +1,6 @@
 package com.nhl.dflib.series;
 
 import com.nhl.dflib.DataFrame;
-import com.nhl.dflib.Index;
 import com.nhl.dflib.Series;
 import com.nhl.dflib.unit.SeriesAsserts;
 import org.junit.Test;
@@ -9,7 +8,7 @@ import org.junit.Test;
 public class ByRowSeries_HeadTest {
 
     private <T> ByRowSeries<T> createSeries(T... data) {
-        return new ByRowSeries<>(DataFrame.forSequenceFoldByRow(Index.forLabels("a", "b"), data));
+        return new ByRowSeries<>(DataFrame.newFrame("a", "b").foldByRow(data));
     }
 
     @Test
@@ -26,7 +25,7 @@ public class ByRowSeries_HeadTest {
 
     @Test
     public void test_OutOfBounds() {
-        Series<String> s = createSeries("a", "b", "c").head(4);
-        new SeriesAsserts(s).expectData("a", "b", "c");
+        Series<String> s = createSeries("a", "b", "c", "d").head(5);
+        new SeriesAsserts(s).expectData("a", "b", "c", "d");
     }
 }