+++ dflib/src/test/java/com/nhl/dflib/unit/BooleanSeriesAsserts.java
@@ -17,7 +17,7 @@ public class BooleanSeriesAsserts {
 
     public BooleanSeriesAsserts expectData(boolean... expectedValues) {
 
-        assertEquals("Unexpected Series length", expectedValues.length, data.length);
+        assertEquals("Unexpected BooleanSeries length", expectedValues.length, data.length);
 
         for (int i = 0; i < expectedValues.length; i++) {