+++ dflib/src/test/java/com/nhl/dflib/unit/DoubleSeriesAsserts.java
@@ -17,7 +17,7 @@ public class DoubleSeriesAsserts {
 
     public DoubleSeriesAsserts expectData(double... expectedValues) {
 
-        assertEquals("Unexpected Series length", expectedValues.length, data.length);
+        assertEquals("Unexpected DoubleSeries length", expectedValues.length, data.length);
 
         for (int i = 0; i < expectedValues.length; i++) {