+++ dflib/src/test/java/com/nhl/dflib/unit/SeriesAsserts.java
@@ -17,6 +17,8 @@ public class SeriesAsserts {
 
     public SeriesAsserts expectData(Object... expectedValues) {
 
+        assertEquals("Unexpected Series length", expectedValues.length, data.length);
+
         for (int i = 0; i < expectedValues.length; i++) {
 
             Object a = data[i];