+++ RELEASE-NOTES.md
@@ -37,6 +37,7 @@
 * #59 Series.valueCount() - utility method for data exploration
 * #60 dflib-jdbc: JdbcConnector to support external transaction management
 * #62 Series.filter(ValuePredicate), Series.filter(BooleanSeries) 
+* #63 Basic `Series.groupBy(..)` 
 
 ## Release 0.5